Update README.md
Stats 507 surveys the software tools that are currently popular among data scientists in academia and industry. The course begins with an accelerated introduction to programming in Python. Next, we focus on Python’s scientific computing stack: numpy, scipy, pandas, and scikit-learn. We also cover using git and some general debugging and profiling techniques. The final part of the course is an introduction to deep learning using PyTorch.

purpose of the repository is to store the code for every homework